SIG-1503 refactor tests incoming messages
